ACS-2290 Improve version validation to include hotfix version too. (#2174)
* ACS-2290 Improve version validation to include hotfix version too.

* ACS-2290 Further improvements to version validation.

* ACS-2290 Fix typo in message.

+++ scripts/travis/build.sh
@@ -65,14 +65,29 @@ else
   pullUpstreamTag "${COM_UPSTREAM_REPO}" "${COM_DEPENDENCY_VERSION}"
 fi
 
-# Ensure that alfresco-community-repo is targetting this version of ACS
+# Ensure that the repository is targeting this version of ACS.
 pushd "$(dirname "${BASH_SOURCE[0]}")/../../../alfresco-community-repo"
 ACS_VERSION_IN_COMMUNITY_REPO="$(retrievePomProperty "acs.version.major").$(retrievePomProperty "acs.version.minor").$(retrievePomProperty "acs.version.revision")"
 popd
-ACS_VERSION=$(mvn -q -Dexec.executable=echo -Dexec.args='${project.version}' --non-recursive exec:exec 2>/dev/null)
-if [[ ${ACS_VERSION} != ${ACS_VERSION_IN_COMMUNITY_REPO}* ]]
+# Get the version properties from enterprise repo and compare with community repo.
+pushd "$(dirname "${BASH_SOURCE[0]}")/../../../alfresco-enterprise-repo"
+MAJOR_FROM_ENTERPRISE=$(mvn -q -Dexec.executable=echo -Dexec.args='${acs.version.major}' --non-recursive exec:exec 2>/dev/null)
+MINOR_FROM_ENTERPRISE=$(mvn -q -Dexec.executable=echo -Dexec.args='${acs.version.minor}' --non-recursive exec:exec 2>/dev/null)
+REVISION_FROM_ENTERPRISE=$(mvn -q -Dexec.executable=echo -Dexec.args='${acs.version.revision}' --non-recursive exec:exec 2>/dev/null)
+if [[ "${MAJOR_FROM_ENTERPRISE}.${MINOR_FROM_ENTERPRISE}.${REVISION_FROM_ENTERPRISE}" != ${ACS_VERSION_IN_COMMUNITY_REPO} ]]
 then
-    printf "Referenced version of community repo specifies ${ACS_VERSION_IN_COMMUNITY_REPO} in pom.xml, but this is ${ACS_VERSION}."
+    printf "Referenced version of community repo specifies \"${ACS_VERSION_IN_COMMUNITY_REPO}\" in pom.xml but enterprise repo specifies \"${MAJOR_FROM_ENTERPRISE}.${MINOR_FROM_ENTERPRISE}.${REVISION_FROM_ENTERPRISE}\"."
+    exit 1
+fi
+# Get the hotfix label (including a dot) from alfresco-enterprise-repo, or an empty string if it's not set.
+HOTFIX_LABEL_IN_ENTERPRISE_REPO="$(retrievePomProperty "acs.version.label")"
+popd
+ACS_VERSION=$(mvn -q -Dexec.executable=echo -Dexec.args='${project.version}' --non-recursive exec:exec 2>/dev/null)
+# Create a regular expression from the repo properties.
+VERSION_REGEX="^${ACS_VERSION_IN_COMMUNITY_REPO}${HOTFIX_LABEL_IN_ENTERPRISE_REPO}([^.0-9].*)?$"
+if ! [[ ${ACS_VERSION} =~ ${VERSION_REGEX} ]]
+then
+    printf "Referenced version of community repo specifies \"${ACS_VERSION_IN_COMMUNITY_REPO}\" in pom.xml and enterprise repo specifies \"${HOTFIX_LABEL_IN_ENTERPRISE_REPO}\", but this is ${ACS_VERSION}."
     exit 1
 fi
 
@@ -106,6 +121,31 @@ if [[ "${SHARE_DEPENDENCY_VERSION}" =~ ^.+-SNAPSHOT$ ]] ; then
 else
   pullUpstreamTagAndBuildDockerImage "${SHARE_UPSTREAM_REPO}" "${SHARE_DEPENDENCY_VERSION}" "-Pbuild-docker-images -Pags -Dlicense.failOnNotUptodateHeader=true -Ddocker.quay-expires.value=NEVER -Ddependency.alfresco-community-repo.version=${COM_DEPENDENCY_VERSION} -Ddependency.alfresco-enterprise-repo.version=${ENT_DEPENDENCY_VERSION}"
 fi
+
+pushd "$(dirname "${BASH_SOURCE[0]}")/../../../alfresco-enterprise-share"
+MAJOR_FROM_SHARE=$(mvn -q -Dexec.executable=echo -Dexec.args='${version.major}' --non-recursive exec:exec 2>/dev/null)
+MINOR_FROM_SHARE=$(mvn -q -Dexec.executable=echo -Dexec.args='${version.minor}' --non-recursive exec:exec 2>/dev/null)
+REVISION_FROM_SHARE=$(mvn -q -Dexec.executable=echo -Dexec.args='${version.revision}' --non-recursive exec:exec 2>/dev/null)
+if [[ "${MAJOR_FROM_SHARE}.${MINOR_FROM_SHARE}.${REVISION_FROM_SHARE}" != ${ACS_VERSION_IN_COMMUNITY_REPO} ]]
+then
+    printf "Referenced version of community repo specifies \"${ACS_VERSION_IN_COMMUNITY_REPO}\" in pom.xml but enterprise share specifies \"${MAJOR_FROM_SHARE}.${MINOR_FROM_SHARE}.${REVISION_FROM_SHARE}\"."
+    # This is broken at the moment - uncomment this when Share is fixed.
+    #exit 1
+fi
+# Check the major part of the Share dependency versions match those from acs-packaging.
+COM_DEP_VERSION_FROM_SHARE="$(retrievePomProperty "dependency.alfresco-enterprise-repo.version")"
+if [[ $(echo ${COM_DEP_VERSION_FROM_SHARE} | cut -d "." -f 1) != $(echo ${COM_DEPENDENCY_VERSION} | cut -d "." -f 1) ]]
+then
+    printf "Community repo version from Share (${COM_DEP_VERSION_FROM_SHARE}) isn't similar to version from acs-packaging (${COM_DEPENDENCY_VERSION})."
+    exit 1
+fi
+ENT_DEP_VERSION_FROM_SHARE="$(retrievePomProperty "dependency.alfresco-community-repo.version")"
+if [[ $(echo ${ENT_DEP_VERSION_FROM_SHARE} | cut -d "." -f 1) != $(echo ${ENT_DEPENDENCY_VERSION} | cut -d "." -f 1) ]]
+then
+    printf "Enterprise repo version from Share (${ENT_DEP_VERSION_FROM_SHARE}) isn't similar to version from acs-packaging (${ENT_DEPENDENCY_VERSION})."
+    exit 1
+fi
+popd
 
 # Build the current project
 mvn -B -ntp -V -q install -DskipTests -Dmaven.javadoc.skip=true -Pbuild-docker-images -Pags ${REPO_IMAGE} ${SHARE_IMAGE}
